obsidian_scripts/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
commonmark==0.9.1
DateTime==5.5
fasteners==0.19
idna==3.10
Levenshtein==0.26.1
load-dotenv==0.1.0
logging==0.4.9.6
path==17.1.0
Pygments==2.19.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
ratio==0.4.0
requests==2.32.3
rich==12.6.0
setuptools==75.8.0
urllib3==2.3.0
watchdog==6.0.0
zope.interface==7.2
langdetect==1.0.9
numpy==2.2.4
scikit-learn==1.6.1
